// src/shared/enums/events.enum.ts
export enum Events {
    CommitCommentEvent = 'CommitCommentEvent',
    CreateEvent = 'CreateEvent',
    DeleteEvent = 'DeleteEvent',
    ForkEvent = 'ForkEvent',
    GollumEvent = 'GollumEvent',
    IssueCommentEvent = 'IssueCommentEvent',
    IssuesEvent = 'IssuesEvent',
    MemberEvent = 'MemberEvent',
    PublicEvent = 'PublicEvent',
    PullRequestEvent = 'PullRequestEvent',
    PullRequestReviewEvent = 'PullRequestReviewEvent',
    PullRequestReviewCommentEvent = 'PullRequestReviewCommentEvent',
    PullRequestReviewThreadEvent = 'PullRequestReviewThreadEvent',
    PushEvent = 'PushEvent',
    ReleaseEvent = 'ReleaseEvent',
    SponsorshipEvent = 'SponsorshipEvent',
    WatchEvent = 'WatchEvent'
}